remove print and change name interaction2 to interaction
