+++ src/control/interaction.py
@@ -6,23 +6,34 @@ from queue import Queue
 import rospy
 from std_srvs.srv import Empty
 from utils.utils import add_masks_to_image
+
 class interaction_control:
     def __init__(self, sam=None, image=None):
         self.sam = sam
         self.image = (
             image if image is not None else np.zeros((512, 512, 3), dtype=np.uint8)
         )
-        self.boxes = []
-        self.masks = None
         self.tk_root = None
-        self.image_queue = Queue()
-        self.anchors = []
         self.rect = None
+        self.image_queue = Queue()
+        # Inputs
+        self.boxes = None        
+        self.anchors = []
+        self.anchors_label = []
+        self.current_label = 1
+
+        # Sam
+        self.num_masks = 1
+        self.masks_array = []
+        self.masks = []
 
     def on_click(self, event=None):
         self.init_x = event.x
         self.init_y = event.y
-        self.rect = None
+        if self.rect is not None:
+            self.canvas.delete(self.rect)
+            self.rect = None
+            self.boxes = None
 
     def on_release(self, event=None):
         self.end_x = event.x
@@ -30,7 +41,7 @@ class interaction_control:
         self.push_box()
 
     def on_move_press(self, event):
-        if self.rect:
+        if self.rect is not None:
             if event.x < 0:
                 end_x = 0
             else:
@@ -46,89 +57,159 @@ class interaction_control:
     def push_box(self):
         ori_init_x, ori_init_y = self.convert_coodinates(self.init_x, self.init_y)
         ori_end_x, ori_end_y = self.convert_coodinates(self.end_x, self.end_y)
-        self.boxes.append([ori_init_x, ori_init_y, ori_end_x, ori_end_y])
+        self.boxes = np.array([ori_init_x, ori_init_y, ori_end_x, ori_end_y])
 
     def on_right_click(self, event=None):
         if event:
             self.right_click_x = event.x
             self.right_click_y = event.y
             ori_right_x, ori_right_y = self.convert_coodinates(self.right_click_x,self.right_click_y)
-            self.anchors.append(np.array([[ori_right_x, ori_right_y]]))
+            self.anchors.append([ori_right_x, ori_right_y])
+            self.anchors_label.append(self.current_label)
             
-            circle_radius = 5
-            self.canvas.create_oval(
-                self.right_click_x - circle_radius,
-                self.right_click_y - circle_radius,
-                self.right_click_x + circle_radius,
-                self.right_click_y + circle_radius,
-                fill='red'
-            )
+            if self.current_label == 1:
+                circle_radius = 5
+                self.canvas.create_oval(
+                    self.right_click_x - circle_radius,
+                    self.right_click_y - circle_radius,
+                    self.right_click_x + circle_radius,
+                    self.right_click_y + circle_radius,
+                    fill='blue'
+                )
+            else:
+                cross_size = 5
+                width = 2
+                self.canvas.create_line(
+                    self.right_click_x - cross_size,
+                    self.right_click_y - cross_size,
+                    self.right_click_x + cross_size,
+                    self.right_click_y + cross_size,
+                    fill='red',
+                    width=width
+                )
+                self.canvas.create_line(
+                    self.right_click_x + cross_size,
+                    self.right_click_y - cross_size,
+                    self.right_click_x - cross_size,
+                    self.right_click_y + cross_size,
+                    fill='red',
+                    width=width
+                )
 
-    def predict_by_boxes(self):
+    def toggle_label(self):
+        self.current_label = 1 - self.current_label  # Switch between 0 and 1
+        if self.current_label == 1:
+            self.label_button.config(text="Foreground", bg="blue", fg="white",
+                                    activebackground="darkblue", activeforeground="white")
+        else:
+            self.label_button.config(text="Background", bg="red", fg="white",
+                                    activebackground="darkred", activeforeground="white")
+
+    def predict_by_prompts(self):
         image_np = np.array(self.image)
         self.sam.set_image(image_np)
-        masks_array = []  # Initialize an empty list to store the mask images
-        if len(self.boxes) == 1:
-            trans_boxes = np.array(self.boxes[0])
-            masks, _, _ = self.sam.predict(
-                point_coords=None,
-                point_labels=None,
-                box=trans_boxes[None, :],
-                multimask_output=False,
-            )
-            mask_image = (masks[0] * 255).astype(np.uint8)
-            masks_array.append(mask_image)  # Append the mask image to masks_array
-            masks = torch.from_numpy(masks).unsqueeze(0)
-            Image.fromarray(mask_image).save(f"outputs/images/mask_to_track0.jpg")           
-            print(f"{len(self.boxes)} mask generated")
-        elif len(self.boxes) > 1:
-            maskss = []
-            if len(self.anchors) == len(self.boxes):
-                for i, i_box in enumerate(self.boxes)                                     :
-                    self.sam.set_image(image_np)
-                    trans_boxes = np.array(i_box)
-                    masks, _, _ = self.sam.predict(
-                        point_coords=self.anchors[i],
-                        point_labels=np.array([1]),
-                        box=trans_boxes[None, :],
-                        multimask_output=True,
-                    )
-                    maskss.append(torch.from_numpy(masks[0]).unsqueeze(0))
-                    for i, msss in enumerate(maskss):
-                        print(i, msss.size())
-                    mask_image = (masks[0] * 255).astype(np.uint8)
-                    masks_array.append(mask_image)  # Append the mask image to masks_array
-                    Image.fromarray(mask_image).save(f"outputs/images/mask_to_track{i}.jpg")
-                print(f"{len(self.boxes)} masks generated")
+        anchors_np = self.safe_np_array(self.anchors)
+        anchors_label_np = self.safe_np_array(self.anchors_label)
+        mask, _, _ = self.sam.predict(
+            point_coords=anchors_np,
+            point_labels=anchors_label_np,
+            box=self.boxes,
+            multimask_output=False,
+        )
+        mask_image = (mask[0] * 255).astype(np.uint8)
+        mask = torch.from_numpy(mask[0]).unsqueeze(0)
+        print(f"Mask{self.num_masks} generated")
+        return mask, mask_image
 
-            else:
-                raise ValueError("Error: The number of boxes and anchors are unmatched.")    
-            masks = torch.stack(maskss, dim=0)
+    def safe_np_array(self, x):
+        if x == []:
+            x_np = None
         else:
-            raise ValueError("No boxes drawn")
-        return_masks = masks.to(self.sam.device)
-        return return_masks, masks_array
-
-    def predict_masks(self):
-        self.masks, mask_images = self.predict_by_boxes()
+            x_np = np.array(x)
+        return x_np
+    
+    def predict_mask(self):
+        mask, mask_image = self.predict_by_prompts()
+        self.masks.append(mask)
+        self.masks_array.append(mask_image)
         # Display the masked image
-        if len(mask_images) > 0:
-            original_image = np.array(self.image)
-            masked_image = add_masks_to_image(original_image, mask_images, bgr=False)            
-            masked_image_pil = Image.fromarray(masked_image)
-            self.mask_photo = ImageTk.PhotoImage(self.resize_image(masked_image_pil))
+        current_image_np = np.array(self.image)
+        masked_image = add_masks_to_image(current_image_np, self.masks_array,bgr=False)        
+        masked_image_pil = Image.fromarray(masked_image)
+        self.mask_photo = ImageTk.PhotoImage(self.resize_image(masked_image_pil))
+        self.canvas.itemconfig(self.image_id, image=self.mask_photo)
+        self.canvas.update()
+        # Clear inputs and remove drawn objects
+        self.boxes = None
+        self.anchors = []
+        self.anchors_label = []
+        if self.rect is not None:
+            self.canvas.delete(self.rect)
+            self.rect = None
 
-            self.canvas.itemconfig(self.image_id, image=self.mask_photo)
-            self.canvas.update()
+        # Remove all drawn circles and crosses
+        drawn_objects = self.canvas.find_all()
+        for obj in drawn_objects:
+            if obj != self.image_id:
+                self.canvas.delete(obj)
+
+        self.canvas.update()
+        self.num_masks += 1
     
     def reset(self):
         self.canvas.delete("all")
+        # Reset Inputs
         self.rect = None        
-        self.boxes = []
+        self.boxes = None
         self.anchors = []
-        self.masks = None
+        self.anchors_label = []
+
+        # Reset sam
+        self.num_masks = 1
+        self.masks_array = []
+        self.masks = []
         self.mask_photo = None
+
+        # Reset Image
         self.image_id = self.canvas.create_image(0, 0, anchor=tk.NW, image=self.photo)
+        self.canvas.update()
+        print("All Cleared")
+
+    def reselect(self):
+        # Check if selection is done.
+        if self.boxes is not None or self.anchors!=[] or self.anchors_label!=[]:
+            if self.rect is not None:
+                self.canvas.delete(self.rect)
+                self.rect = None
+
+            # Remove all drawn circles and crosses
+            drawn_objects = self.canvas.find_all()
+            for obj in drawn_objects:
+                if obj != self.image_id:
+                    self.canvas.delete(obj)            
+            # Reset Inputs
+            self.boxes = None
+            self.anchors = []
+            self.anchors_label = []
+        else:
+            # Remove latest output
+            # Display previous version
+            if len(self.masks_array) > 0:
+                self.masks_array.pop(-1)
+                self.masks.pop(-1)
+                self.num_masks -= 1
+                if len(self.masks_array) == 0:
+                    self.mask_photo = self.photo
+                else:                
+                    current_image_np = np.array(self.image)
+                    masked_image = add_masks_to_image(current_image_np, self.masks_array,bgr=False)        
+                    masked_image_pil = Image.fromarray(masked_image)
+                    self.mask_photo = ImageTk.PhotoImage(self.resize_image(masked_image_pil))
+            else:
+                # All mask cleared
+                self.mask_photo = self.photo
+            self.image_id = self.canvas.create_image(0, 0, anchor=tk.NW, image=self.mask_photo)
+        print("Removed")        
         self.canvas.update()
         
     def tracking(self):
@@ -154,7 +235,7 @@ class interaction_control:
             original_y = photo_y*self.ratio_y
             return original_x, original_y
         else:
-            return photo_x, photo_y
+            return photo_x, photo_y                       
 
     def do_interact(self, image=None, tk_root=None, ros=False):
         if tk_root:  # If a new Tk root is provided
@@ -179,28 +260,48 @@ class interaction_control:
                 self.canvas.bind("<Button 3>", self.on_right_click)
 
             self.image_id = self.canvas.create_image(0, 0, anchor=tk.NW, image=self.photo)  # Display original image
-        button_frame = tk.Frame(self.tk_root)
-        button_frame.pack(pady=10)
+        button_frame_top = tk.Frame(self.tk_root)
+        button_frame_top.pack(pady=5)
+
+        button_frame_bottom = tk.Frame(self.tk_root)
+        button_frame_bottom.pack(pady=5)
+
         button_bg_color = "#616161"
         button_fg_color = "#FFFFFF"
         button_active_bg_color = "#424242"
         button_active_fg_color = "#FFFFFF"
 
-        mask_button = tk.Button(button_frame, text="Generate Masks", command=self.predict_masks,
+        label_button = tk.Button(button_frame_top, text="Foreground", command=self.toggle_label,
+                                width=15, height=2, font=("Helvetica", 12), bg="blue", fg="white",
+                                activebackground="darkblue", activeforeground="white")
+        label_button.pack(side=tk.LEFT, padx=10, pady=5)
+        self.label_button = label_button
+
+        mask_button = tk.Button(button_frame_top, text="Generate Mask", command=self.predict_mask,
                                 width=15, height=2, font=("Helvetica", 12), bg=button_bg_color, fg=button_fg_color,
                                 activebackground=button_active_bg_color, activeforeground=button_active_fg_color)
         mask_button.pack(side=tk.LEFT, padx=10, pady=5)
-
-        reset_button = tk.Button(button_frame, text="Reselect", command=self.reset,
-                                width=15, height=2, font=("Helvetica", 12), bg=button_bg_color, fg=button_fg_color,
-                                activebackground=button_active_bg_color, activeforeground=button_active_fg_color)
-        reset_button.pack(side=tk.LEFT, padx=10, pady=5)
+        self.mask_button = mask_button
 
         self.track_pressed = 0
-        track_button = tk.Button(button_frame, text="Start Tracking", command=self.tracking,
+        track_button = tk.Button(button_frame_top, text="Start Tracking", command=self.tracking,
                                 width=15, height=2, font=("Helvetica", 12), bg="green", fg="white",
                                 activebackground="darkgreen", activeforeground="white")
         track_button.pack(side=tk.LEFT, padx=10, pady=5)
+        self.track_button = track_button
+
+        reselect_button = tk.Button(button_frame_bottom, text="Reselect", command=self.reselect,
+                                    width=15, height=2, font=("Helvetica", 12), bg=button_bg_color, fg=button_fg_color,
+                                    activebackground=button_active_bg_color, activeforeground=button_active_fg_color)
+        reselect_button.pack(side=tk.LEFT, padx=10, pady=5)
+        self.reselect_button = reselect_button
+
+        reset_button = tk.Button(button_frame_bottom, text="Clear All", command=self.reset,
+                                width=15, height=2, font=("Helvetica", 12), bg=button_bg_color, fg=button_fg_color,
+                                activebackground=button_active_bg_color, activeforeground=button_active_fg_color)
+        reset_button.pack(side=tk.LEFT, padx=10, pady=5)
+        self.reset_button = reset_button
+        
         while not self.track_pressed:
             self.tk_root.update()
             self.tk_root.after(10)
@@ -212,7 +313,8 @@ class interaction_control:
                 start_inference()
             except rospy.ServiceException as e:
                 print("Service call failed: %s"%e)
-        print(self.masks.size())
+
+        return_masks = torch.stack(self.masks, dim=0)
+        return_masks = return_masks.to(self.sam.device)
         print("Start tracking")
-        assert False
-        return self.masks
+        return return_masks
